updates
- allow dry run
- assume current dir
- improve print output

+++ main.go
@@ -11,15 +11,21 @@ import (
 	"strings"
 )
 
-func renameImageFile(filePath, newFileName string, index, numDigits int) error {
+func renameImageFile(filePath, newFileName string, index, numDigits int, dryRun bool) error {
 	ext := strings.ToLower(filepath.Ext(filePath))
 	formatString := "%0" + strconv.Itoa(numDigits) + "d"
 	newFilePath := filepath.Join(filepath.Dir(filePath), strings.ToLower(newFileName)+"-"+fmt.Sprintf(formatString, index)+ext)
+
+	if dryRun {
+		fmt.Printf("Dry run: renaming '%s' to '%s'\n", filePath, newFilePath)
+		return nil
+	}
 
 	err := os.Rename(filePath, newFilePath)
 	if err != nil {
 		return err
 	}
+	fmt.Printf("Renamed file '%s' to '%s'\n", filePath, newFilePath)
 	return nil
 }
 
@@ -39,15 +45,25 @@ func (d byDate) Less(i, j int) bool {
 
 func main() {
 	order := flag.String("order", "name", "Specify the order of the input files (name or date)")
+	dryRun := flag.Bool("d", false, "Dry run: print the proposed changes but do not perform them")
 	flag.Parse()
 
 	if len(flag.Args()) < 2 {
-		fmt.Println("Usage: sequentialize [-order=name|date] directory new_file_name")
+		fmt.Println("Usage: sequentialize [-order=name|date] [-d] directory new_file_name")
 		return
 	}
 
 	dir := flag.Arg(0)
 	newFileName := flag.Arg(1)
+
+	if dir == "." {
+		var err error
+		dir, err = os.Getwd()
+		if err != nil {
+			fmt.Println("Error getting current directory:", err)
+			return
+		}
+	}
 
 	files, err := ioutil.ReadDir(dir)
 	if err != nil {
@@ -66,22 +82,16 @@ func main() {
 		return
 	}
 
-	imageFiles := []string{}
-	for _, file := range files {
-		ext := strings.ToLower(filepath.Ext(file.Name()))
-		if ext == ".jpg" || ext == ".jpeg" || ext == ".png" || ext == ".heic" {
-			imageFiles = append(imageFiles, filepath.Join(dir, file.Name()))
+	numDigits := len(fmt.Sprintf("%d", len(files)))
+
+	for i, file := range files {
+		if file.IsDir() {
+			continue
 		}
-	}
-
-	numDigits := len(fmt.Sprintf("%d", len(imageFiles)))
-
-	for i, filePath := range imageFiles {
-		err := renameImageFile(filePath, newFileName, i+1, numDigits)
+		filePath := filepath.Join(dir, file.Name())
+		err := renameImageFile(filePath, newFileName, i+1, numDigits, *dryRun)
 		if err != nil {
 			fmt.Printf("Error renaming file '%s': %v\n", filePath, err)
-		} else {
-			fmt.Printf("Renamed file '%s'\n", filePath)
 		}
 	}
 }
